feat: add accordion styling for mobile
