+++ src/app/pages/FAQ/constants.ts
@@ -2,41 +2,49 @@ export const FAQContent = [
   {
     id: 'item-1',
     header: 'Who is eligible?',
-    content: 'Lorem ipsum dolor sit amet.'
+    content:
+      'Lorem ipsum dolor sit amet consectetur adipisicing elit. Corporis inventore ex, laborum alias iure, autem pariatur sit perferendis fugit porro id sapiente adipisci esse fuga!'
   },
   {
     id: 'item-2',
     header: 'What is the time commitment?',
-    content: 'Lorem ipsum dolor sit amet.'
+    content:
+      'Lorem ipsum dolor sit amet consectetur adipisicing elit. Corporis inventore ex, laborum alias iure, autem pariatur sit perferendis fugit porro id sapiente adipisci esse fuga!'
   },
   {
     id: 'item-3',
     header: 'How many people will be accepted?',
-    content: 'Lorem ipsum dolor sit amet.'
+    content:
+      'Lorem ipsum dolor sit amet consectetur adipisicing elit. Corporis inventore ex, laborum alias iure, autem pariatur sit perferendis fugit porro id sapiente adipisci esse fuga!'
   },
   {
     id: 'item-4',
     header: 'How will UP-Graders be selected?',
-    content: 'Lorem ipsum dolor sit amet.'
+    content:
+      'Lorem ipsum dolor sit amet consectetur adipisicing elit. Corporis inventore ex, laborum alias iure, autem pariatur sit perferendis fugit porro id sapiente adipisci esse fuga!'
   },
   {
     id: 'item-5',
     header: 'Who will my mentors be?',
-    content: 'Lorem ipsum dolor sit amet.'
+    content:
+      'Lorem ipsum dolor sit amet consectetur adipisicing elit. Corporis inventore ex, laborum alias iure, autem pariatur sit perferendis fugit porro id sapiente adipisci esse fuga!'
   },
   {
     id: 'item-6',
     header: 'Can I request teammates?',
-    content: 'Lorem ipsum dolor sit amet.'
+    content:
+      'Lorem ipsum dolor sit amet consectetur adipisicing elit. Corporis inventore ex, laborum alias iure, autem pariatur sit perferendis fugit porro id sapiente adipisci esse fuga!'
   },
   {
     id: 'item-7',
     header: 'What roles are available?',
-    content: 'Lorem ipsum dolor sit amet.'
+    content:
+      'Lorem ipsum dolor sit amet consectetur adipisicing elit. Corporis inventore ex, laborum alias iure, autem pariatur sit perferendis fugit porro id sapiente adipisci esse fuga!'
   },
   {
     id: 'item-8',
     header: 'When is the application due?',
-    content: 'Lorem ipsum dolor sit amet.'
+    content:
+      'Lorem ipsum dolor sit amet consectetur adipisicing elit. Corporis inventore ex, laborum alias iure, autem pariatur sit perferendis fugit porro id sapiente adipisci esse fuga!'
   }
 ];
